requirements.txt
twilio==8.12.0
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0 